*** empty log message ***
git-svn-id: svn://localhost/svn/uclinux-dist/trunk@287 3e2a5524-39e8-4cb9-84a9-c49da69cba3a

+++ linux-2.6.x/include/asm-blackfin/dpmc.h
@@ -35,9 +35,9 @@
 
 
 int transit_to_newmode (int newmode);
-int calc_volt(void);
+unsigned long calc_volt(void);
 int calc_vlev(int vlt);
-double change_voltage(double volt);
+unsigned long change_voltage(unsigned long volt);
 int calc_msel(int vco_hz);
 unsigned long change_frequency(unsigned long vco_mhz);
 int set_pll_div(unsigned short sel,unsigned char flag);
